Fix label changing in the templateObjectCell
The event from the textEditorField was being handled twice; once
because the templateObjectCell observed the textEditorField itself
and again because it observed the actual parent component of the
textEditorField for "action" as well.

Stopping propagation seemed pretty sane, though I always worry that
will only introduce a bug of a different sort where we won't
receive the event we expect in some other context.

Anyway, the double handling caused the event to trigger saving the
newLabel twice; the second application failed which resulted in the
change being "rejected." This is what caused the value presented to be
reverted even though the new label was successfully applied.