Hide the montageId field whenever possible
I couldn't use a condition because its introduction caused problems with
the flex box layout specified by the parent, so I stole the same
classname.

Anyway now we only show the segment when there is no component or when
there is a component and the montageId is not the same as the 
component's label.