Add application to library and ensure it works
This makes it possible to get add a reference to the application
inside a template if needed.

It also introduces the idea of userObjects to the reelProxy which
will squelch the serialization of an object's prototype entry
in the declaration.

There are further wrinkles to explore but this covers some of the less
adventurous cases of accessing the application from within a
a declaration.