Coalesce multiple uploads on the same fileCell
Starting upload batch A and then introducing batch B while A is pending
will now be represented by the fileCell as progress pending on the
combined A+B uploads.

We're not doing anything to stagger them at this time. We could also
simply reject uploads while in the middle of uploading, but I like how
unexpected this sort of accommodation is, especially if we have
a nice progress indicator to go with it.

Generally, uploading is very likely to kill the web socket connection;
we have some work throttle this in the environment/service layer.