Improve accuracy of bytes uploaded
Without some smoothing it's not a particularly great measure for
presenting to the user; while accurate about bytes it doesn't account
for latency/actually transmitting that final leg from client to
server.

It will actually finish too early as the bytes being reported are bytes
being read by the browser, not bytes transferred to our backend.