Add project-level utility to load moduleBlueprints
This gives us a central place for loading blueprints using the correct
require and non-deprecated API.

As noted this should eventually be moved into a document or something
representing the package itself and the editing API for the package.

Right now the projectController has taken on these responsibilities
in addition to its main responsibility of loading a project.