Ignore drag leaves that don't correspond to enters
It was possible to force the leave into the extreme negative territory 
by dragging in something that we would ignore when entering such as
an object reference from within filament itself.

This would make it impossible to indicate acceptance of a valid drag
as the next drag enter would not bring the hover count above 0.