Restore dragging and dropping throughout filament
We don't need to stop the defaults unless the drag is for a file
that would trigger the default behavior of navigating to the dropped
file.

Unchecked, this was preventing us from dragging and dropping text/plain
throughout the application into textfields that relied on the
default behavior.

Generally, we probably will want to also have some more specific
handling in these cases, but right now we do expect the default to
happen.