Remove extension panel
Before I forget to take it out. I can't wait to formalize the extensions
and offer the system for others to use, but it's not ready for that
sort of exposure just yet.

I'm leaving the UI in place for the time being as it's not a horrible
starting point for at least listing the extensions.