Remove manual committing from client side
While we need access to read branch status for reporting and making
various parts of the interface available, I'm keeping the burden of
committing and pushing on the environment side alone.

This is a bit of a moment of clarity I've had about this workflow.

I haven't done anything yet to split up the large chain of commits we
generate in the course of saving, but that's next up after I verify
that commits do in fact end up on the right branch and get pushed.

But otherwise, heavy lifting on the environment, higher-level
abstraction operations on the client side.

So while the client side might only operate with a verb like "save",
the environment is free to fulfill that higher level operation
however it needs to.

The client is then free to read minimal information back from the
environment to customize UI accordingly.

I still intend to centralize most of the project-wide editing
operations into the projectDocument.