Start moving module creation to projectDocument
Don't have time to move everything, and I'd like to put the shadow
updating on the environment so the project document is yet more
environment agnostic.

Anyway, I expect more of the creation to move over into
projectDocument as we reduce the project controller's role in these
sorts of things.