Move modalEditor concern to componentEditor
The modal editor is only modal for the editor that requested it.
Work needs to be done to make sure that document switching still works
as expected; the same editor should technically be able to 
switch between documents that are either showing modal editor A,
modal editor B, or showing no modal editor at all.