Improve project bootstrapping progress reporting
Foremost, we now start initializing a project immediately after we
have created the repository.

When landing on a projectUrl that is currently initializing we now
report as such, avoiding the need to manually trigger project creation
that is currently underway.

For completeness can now also trigger creating new repositories
when directly browsing to a projectUrl for a repository that does
not exist.

Error reporting should also be slightly improved as we now detect a
variety of other scenarios that you might end up in and report
more specific errors to the user.