Stop considering doc closing when close cancelled
Mainly, we needed to make sure we removed the closingPromise from the 
document closing Promise map when canceling.

In fixing this ended up having to clean up some of this to distill
the problem a bit so those changes are now there. If nothing else
we are no longer creating an unnecessary array each time we close a
document now; we rely on the in scope editing document to know which
document to close instead of passing it along and spreading the promise
for editor and document.