Use a single package require for all documents
Documents now receive a package require for the package url when loading.
The main goal is to avoid creating a new montage sandbox for each
ReelDocument loaded.
Instead of delegating this to the ReelDocument (or any document) it is
now the responsibility of the ProjectController to provide a package
require.