Add serialization to the preload document
By adding a serialization we preload all the css files.
I used the details.reel of popcorn because it is a case that is
realistic and dramatically increases the time to load a new reel.
In my tests showing the component editor UI after the reel document is
ready was taking ~2s. With this preload document this is cut down to a
few hundred ms, it depends on the reel shape it's loading.