Make load a constructor property
With this move we make it possible to create a Document before loading
everything needed to instantiate it.
It allows us to put it into to document list of the ProjectController
and prevent the creation of multiple instances of the same document
when the user clicks several times on the package explorer.