Remove auto-forking tutorial
We're going to feature example projects on the project listing page
instead of dumping a user into the relatively complex popcorn app.

Eventually with work making the app itself more approachable and work
to provide guidance within filament we should be able to reconsider
the forking on first-run.

i.e. The feature should not be considered unwanted, just less than
desirable until the project is up to the task of being something to
which we can point newcomers.