Quickly get the sceneEditor rendering again
This was relying on the stageObject to normalize the path such that
it was resolving on the project domain. Without the stage, we need to
do this ourselves.

I'm pretty certain we have this sort of code somewhere else in a sane
repeatable manner. But I need to get out the door and I'd like this
working in the next deployment.