Allow the delegate the define how the project is loaded
This application delegate has a new function - loadProject - that is
responsible for showing the progress message for loading the project
and ask the user if it wants to try again when loading the project
fails.

Previously the loading message was being handled by the willLoadProject
phase which wasn't the right place to handle the load / retry cycle.