Set a background image for the imageViewer editor
To obscure whatever might be beneath editors should provide a
background. Additionally, this helps attach the image editor document to
the tab by way of color.

Eventually I want to give control over the brightness of this editor so
I didn't match the entire background to the tab, just the toolbar.