Ensure goto overlay preserves correct activeTarget
The overlay was always setting the activeTarget back upon hiding,
we needed to make sure it either doesn't set it or that it sets it to
what we expect.

I'm deprecating the openUrl event dispatching as a means of opening
urls, or at the very least, using the direct API whenever I need to know
when the document has opened.