Implement an undoManager for editorFields
This puts us in charge of undoability within editorFields instead of
relying on the native undo within text inputs.

In addition to putting us a step closer to being able to use the native
undo keyboard shortcut as we're closer to the native expectations, this
also addresses the issue where you could be editing an object you just 
added and then undo its addition, which was…unfortunate.

Now you can't delete the object out from under you by way of undoing
its addition.