Fallback to ::after pseudo class to display shortcuts.
Using the pseudo class does make it possible to go with this approach: http://michaelhue.com/keyscss/ (one box per key to mimic the keyboard keys).
But it does not require as much work. To make it work a new menuItemRow component would need to be created to wrap the label in a flexbox grow 2 and put the shortcut on the side on a grow 0.