Make initial loader opaque to hide app loading
I also brought the loader widget higher up along the z axis so it's
visible if necessary,but it doesn't look quite right as it doesn't have
the same dimensions as the loader that follows it (very quickly) due
to the text and some extra padding.

I don't want to lose the bootstrapping loader but wouldn't mind it 
being different enough that the jump to the actual longer loader is
less jarring.